requirements.txt
# Numerical Computing
numpy
scipy
pandas
arrgh 

# Visualization
matplotlib

# Data Loading


# Deep learning
torch
torchvision